Fixed Icon Bug and Biometrics
Fixed a problem where the icon would not show up on newer devices that used a rounded icon, and added fingerprints to the emulator I am using, making it so that it can recieve fingerprint signals and pass the requirements for a biometric prompt. Now the prompt needs to be added which paves the way for the addition of the map activity.